kismet2rabbit/requirements.txt
requests
scapy-python3
